refactor(mp-2866): move to util/, rename to getGoalSignupCopyVariant
The helper returns a plain string with no reactive behavior, so it
belongs in src/util/ rather than src/composables/, and the `use*`
prefix is misleading. Renamed to `getGoalSignupCopyVariant`.
