+++ src/util/goalSignupCopyVariant.js
